+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Feature] Add screenshot and unlock to screenstreaming
 - [Feature] Rework mfkey32 founded key screen
 - [Feature] Design fix for widget
+- [Feature] Add storage and flipper connected error
 - [CI] Auto git submodule update via gradle task
 - [FIX] Options button in device screen
 - [FIX] WearOS connection
@@ -20,6 +21,7 @@
 - [FIX] Flipper color on error screen
 - [FIX] Fix fatal crash on empty backup
 - [FIX] Widget options not working in some android devices
+- [FIX] Try to fix uncanceled notification
 - [FIX] Move resources from night qualifier
 
 # 1.2.2